SIDM-8932 adapt invitations to work for PIN uplifts
